Removed an errant comment :)
At the top of the code in the previous version, it said it was day 5 code of 100 days of code.  It was *going* to be, but I recycled the code turning into the Virtual Dice app instead.  The real Day 5 project was a password generator, which I'm actually working on right now funnily enough :)